Quick hack to not ignore danish accented letters
The current RHVoice voice implementations completely ignore non-latin
latters (å, æ, ø) as if they're not there. This is a quick test / hack
to at least get them to produce sound.

Since I couldn't find out how to do it in the actual voice, this is just
a quick replacement on the Android level (the only platform where I use
RHVoice, for voice GPS navigation).

It does work, but there's probably a better, lower, layer where this
should be handled.